Remove obsolete reference to xxd.c
The build used to use xxd to generate the source for the compiled in
fonts, and the README references an external implementation for use
on systems that don't have this utility. However, this is now handled
by the included hexdump.c and the external reference is just confusing.

darcs-hash:20080923174500-ec8f4-6b41aea08478c70cc379fcc47cf0fb9de9e62eda.gz