Use the available local to shorten the code.
Ignore-this: e93bf52e655b966769983df3eb231f7e

The cache aging loop already stored the next linked list element in
a local variable, so we can join the list prior to eviction using
that instead of dereferencing the element to be evicted again.

The loop is small enough that I think the referent is clear and
it makes the statements shorter. This version also works if the
eviction happens prior to the list update, as was the case prior
to the recent cleanup.

darcs-hash:20091207195144-cac71-db84816b26d87899597af7b0395e3802dbc526d7.gz